Move all designs to designs update today
